fixup: better positioning, move ai input below options
