+++ packages/compass-components/src/components/generative-ai/generative-ai-input.spec.tsx
